+++ x/emissions/keeper/topic_activation.go
@@ -131,18 +131,26 @@ func (k *Keeper) inactivateTopicWithoutMinWeightReset(ctx context.Context, topic
 	if err != nil {
 		return errors.Wrap(err, "failed to get active topic ids at block")
 	}
+	topicFound := false
 	// Remove the topic from the active topics at the block
 	// If the topic is not found in the active topics at the block, no op
-	newActiveTopicIds := []TopicId{}
 	for i, id := range topicIdsActiveAtBlock.TopicIds {
 		if id == topicId {
-			newActiveTopicIds = append(topicIdsActiveAtBlock.TopicIds[:i], topicIdsActiveAtBlock.TopicIds[i+1:]...)
+			newActiveTopicIds := append(
+				topicIdsActiveAtBlock.TopicIds[:i],
+				topicIdsActiveAtBlock.TopicIds[i+1:]...,
+			)
+			err = k.SetBlockToActiveTopics(ctx, block, types.TopicIds{TopicIds: newActiveTopicIds})
+			if err != nil {
+				return errors.Wrap(err, "failed to set block to active topics")
+			}
+			topicFound = true
 			break
 		}
 	}
-	err = k.SetBlockToActiveTopics(ctx, block, types.TopicIds{TopicIds: newActiveTopicIds})
-	if err != nil {
-		return errors.Wrap(err, "failed to set block to active topics")
+
+	if !topicFound {
+		return errors.Wrap(types.ErrNotFound, "active topic expected to be found in block's active topics list")
 	}
 
 	err = k.topicToNextPossibleChurningBlock.Remove(ctx, topicId)